requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
h11==0.14.0
idna==3.10
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.1
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6